requirements.txt
Django==3.2.7
djangorestframework==3.12.1
djangorestframework-simplejwt
python-dotenv==0.13.0
PyYAML==5.3.1
uritemplate==3.0.1
django-cors-headers==3.5.0
factory_boy==3.2.0
pytest-cov==2.12.0
pytest-mock==3.6.1
pytest==6.2.4
django-filter==2.4.0
psycopg2==2.8.6

